restructure directories and files
