+++ footnote-backend/config/tables.js
